backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.0.1
alembic==1.13.1
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
authlib==1.2.1
httpx==0.25.2
python-multipart==0.0.6
aiofiles==23.2.1
celery==5.3.4
flower==2.0.1
langchain==0.0.350
openai==1.3.7
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
pillow==10.1.0
transformers==4.36.2
torch==2.1.1
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
websockets==10.4
fastapi-websocket-rpc==0.1.25 